+++ tests/testthat/test-do_ViolinPlot.R
@@ -4,6 +4,11 @@ if(base::isFALSE(dep_check[["do_ViolinPlot"]])){
     p <- SCpubr::do_ViolinPlot(sample = sample,
                                features = "EPC1")
     testthat::expect_type(p, "list")
+    
+    p <- SCpubr::do_ViolinPlot(sample = sample,
+                               features = "EPC1",
+                               order = FALSE)
+    testthat::expect_type(p, "list")
   })
 
   testthat::test_that("do_ViolinPlot: PASS - one variable", {
@@ -28,7 +33,8 @@ if(base::isFALSE(dep_check[["do_ViolinPlot"]])){
     p <- SCpubr::do_ViolinPlot(sample = sample,
                                features = "EPC1",
                                split.by = "annotation",
-                               plot_boxplot = FALSE)
+                               plot_boxplot = FALSE,
+                               order = FALSE)
     testthat::expect_type(p, "list")
 
   })
@@ -207,7 +213,8 @@ if(base::isFALSE(dep_check[["do_ViolinPlot"]])){
 
     testthat::expect_error(SCpubr::do_ViolinPlot(sample = sample,
                                                  features = "EPC1",
-                                                 split.by = "orig.ident"))
+                                                 split.by = "orig.ident",
+                                                 order = FALSE))
   })
 
 
@@ -226,13 +233,15 @@ if(base::isFALSE(dep_check[["do_ViolinPlot"]])){
                                features = "EPC1",
                                split.by = "orig.ident",
                                group.by = "annotation",
-                               plot_boxplot = FALSE)
+                               plot_boxplot = FALSE,
+                               order = FALSE)
     testthat::expect_type(p, "list")
 
     p <- SCpubr::do_ViolinPlot(sample = sample,
                                features = "EPC1",
                                split.by = "seurat_clusters",
-                               plot_boxplot = FALSE)
+                               plot_boxplot = FALSE,
+                               order = FALSE)
     testthat::expect_type(p, "list")
   })
 
@@ -306,7 +315,8 @@ if(base::isFALSE(dep_check[["do_ViolinPlot"]])){
 
     testthat::expect_error(SCpubr::do_ViolinPlot(sample = sample,
                                                  features = "EPC1",
-                                                 split.by = "annotation"))
+                                                 split.by = "annotation",
+                                                 order = FALSE))
   })
 }